Revert "feat: Corrige o mapeamento da tabela Cadastra e ajusta tipos de dados…"

+++ sime.api/src/main/java/br/com/sime/api/entities/Feedback.java
@@ -12,7 +12,7 @@ public class Feedback {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_feedback", nullable = false)
-    private Long idFeedback;
+    private Integer idFeedback;
 
     @Column(name = "dt_feedback")
     private LocalDateTime dtFeedback;
